+++ CHANGELOG.md
@@ -3,9 +3,15 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-08-20
+### Added
+- Support for lower-case chars in postcode
+- Some basic unit tests
+
 ## [0.1.0] - 2019-08-19
 ### Added
 - Initial version.
 
-[Unreleased]: https://github.com/codeasone/postcode-spec/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/codeasone/postcode-spec/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/codeasone/postcode-spec/releases/tag/v0.2.0
 [0.1.0]: https://github.com/codeasone/postcode-spec/releases/tag/v0.1.0